Update various usages of `docker-compose` to `docker compose` (#62)
CI broke as of 2024/04/01 because of
[this](https://github.com/actions/runner-images/issues/9557). I update
all of the references to `docker-compose` to `docker compose` as
recommended, which seems to have fixed it. All jobs passed, and at least
some of them ran on the new image (I'm not entirely sure what controls
this; perhaps which runner the job hits?), which tells me that this
change works in both circumstances.