Oplogtoredis: Add configurable Write Parallelism (#70)
This adds a new configuration option, OTR_WRITE_PARALLELISM. This will be an integer that controls how many parallel writer loops deliver the oplog messages to Redis. Each oplog message will be routed to one of the loops based on a SHA256 hash of its database name. The value defaults to 1 if unspecified, which means it parallelism is off by default (since a parallelism of 1 means a single writer, the same as currently).