Correction for objectID issue in last merge (#86)
The last PR merged decode the objectID as either string or objectId
type, but would fail on binary objectIDs. The prior implementation had
converted the objectID to an interface{} type, to be handled downstream,
so this moves back to that approach, while still limiting the
unmarshaling to just the _id field.