Add label to staleness metrics to distinguish published vs duplicate entries (#93)
Oplogtoredis runs two replicas in parallel which both write to redis,
with dedup being done by a lua script. However, both of these replicas
report the latency in time to reach redis, regardless of whether it was
the first to reach redis or not. This means that is one replica is
lagging behind, it will report larger latency values than what’s
actually reaching redis, and we don’t have a way to filter these since
they will be reported at a later time than the earlier successful write
from the other replica.

This adds a return value to the dedup script and uses it as a label for
the latency metrics.

Note: for some reason returning true/false from the lua script was
causing the performance and fault_injection tests to fail (after an
extra long time). Switching to int, returning 0 or 1, worked. Testing
showed that returning false was causing issues, and a search found
[this](https://stackoverflow.com/questions/75403292/lua-script-return-false-is-giving-error-redis-nil-during-the-evalsha)
which seems to align with this issue, where the lua script returning
false is treated as an error, which then caused OTR to retry (until the
dedupe token expired). I [opened an issue with
go-redis](https://github.com/redis/go-redis/issues/3246).