Downgrade benign error log to info (#94)
For commands that aren't transactions, we can safely ignore the missing
applyOps key, since that key will only be there when it is a
transaction. All other commands that do not include the applyOps key
apply only to the admin db, so there is no need to further process
these.