Log and return error for delete denylist that doesn't exist (#96)
When an item in the denylist is deleted, the former behavior didn't
provide any feedback as to whether it existed or not. This PR adds a log
warning and returns 404 for a delete attempt on an entry not in the
denylist.

Additionally, the acceptance test runner was updated to ubuntu 24.04
because 20.04 was deprecated on 4/15.

---------

Co-authored-by: sannivasreddy <94570377+sannivasreddy@users.noreply.github.com>