Upgrade Debian Docker image tags for dev and test containers (#101)
Upgrades the Docker image tags used for Debian-based images in
development and test containers, since versions based on Debian "buster"
no longer run due to offline repositories. As a stopgap, these are
upgraded to the same tool versions but with images based on "bullseye",
since there is a chain of incompatibilities related to an outdated
MongoDB driver that makes further upgrades more difficult. Those will be
addressed in future changes.