Improve precision of staleness metrics using oplog entry wall time (#99)
There are a handful of staleness metrics that compare the timestamp
attached to oplog events with the current time in oplogtoredis. These
originally used the oplog timestamp, since it was already included in
parsed oplog entries, but this is not ideal for metrics since it only
has resolution to the nearest second, with the rest of the timestamp
bits dedicated to ordering instead.

With this change, we use the wall time from each oplog entry instead,
which has resolution to the nearest millisecond.

Now that the metric is more precise, buckets needed to be updated to
account for the increased precision. A set of buckets below 10ms has
been added to reflect typical performance under minimal load. Extra
buckets between 1-10s have also been added, since this area of latency
is perceptually relevant for users expecting to see live changes in
Meteor.

Unit tests have been updated to check that wall time is included in
parsed entries, and existing acceptance tests verify that the parsing
works with real entries. This change was also tested manually to ensure
that the resulting metrics are more accurate and do not average out to
~0.5s staleness, which happened when the precision was only to the
nearest second.