Denylist: Return 204 when non-existent entry is deleted from denylist (#102)
When oplogtoredis receives an HTTP request to delete an entry from the
denylist, if the entry doesn't exist, this
[commit](https://github.com/tulip/oplogtoredis/commit/e1488dbf6af561a0a88c9ffe4a11271d23bb9fb3#diff-245563af9e34754739b12b294f3324c1d42c116ace7ce8735e21dfbc82d8e064)
made it such that the response is a 404. However, this causes issues
with deployment operations at Tulip, where it is valid to send a DELETE
request for a denylist entry even if it was never added. This change
returns to old behavior, where a 204 was returned in these cases.