Retry fetching resume timestamp from redis on startup (#105)
Upon startup, oplogtoredis reads the last processed timestamp from redis
for resuming oplog tailing from that time. However, if there is an
error, it will fall back to the current time, potentially dropping a
portion of the oplog. This PR adds a retry (configurable; default is up
to five times, with a delay from 200ms to 1000ms).

While adding tests for this, I found that the existing tests were
broken, effectively comparing 0==0 and always passing. Fixing these
tests revealed that the fallback to the current time was setting the
starting timestamp to zero, rather than the current time. This is also
fixed in this PR.