Clamp entry staleness metric to zero (#106)
Clock skew can cause the difference between Mongo's reported wall time
and OTR's current time to be a negative value. During a metrics scrape,
if these negative values cause the histogram sum to be negative,
Prometheus will treat it as a metric reset because the sum is otherwise
assumed to be monotonic. As a result, Grafana charts that use the
histogram sum will have artifacts, e.g. large spikes.

This PR addresses the issue by rounding negative values up to 0, as the
skew should only be a few milliseconds at most when using NTP.